# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me/orders/{id}": {
      "get": {
        "operationId": "UsersController_getOrderDetails",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/rooms/{id}/order/{orderId}": {
      "post": {
        "operationId": "RoomsController_linkOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Rooms"
        ]
      }
    },
...
    "/products/{id}": {
      "get": {
        "operationId": "CatalogController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "operationId": "CatalogController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "CatalogController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
...
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_getOrderDetails",
        "summary": "Get order details by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/admin/orders/{id}": {
      "get": {
        "operationId": "OrdersAdminController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/vendors/register": {
      "post": {
        "operationId": "VendorsController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterVendorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Vendors"
        ]
      }
    },
...
    "/upload/image": {
      "post": {
        "operationId": "UploadController_uploadImage",
        "summary": "Upload a single image (optimized to WebP)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "context": {
                    "type": "string",
                    "example": "products"
                  },
                  "contextId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Upload"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/upload/images": {
      "post": {
        "operationId": "UploadController_uploadMultipleImages",
        "summary": "Upload multiple images",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "context": {
                    "type": "string",
                    "example": "products"
                  },
                  "contextId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Upload"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/upload/document": {
      "post": {
        "operationId": "UploadController_uploadDocument",
        "summary": "Upload a document (PDF, Image)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "documentType": {
                    "type": "string",
                    "example": "KYC"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Upload"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/upload/{path}": {
      "delete": {
        "operationId": "UploadController_deleteFile",
        "summary": "Delete a file",
        "parameters": [
          {
            "name": "path",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Upload"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "RISBOW API",
    "description": "\n# RISBOW Ecommerce Super App API\n\nWelcome to the RISBOW API documentation!\n\n## Key Features:\n- 🔐 **Authentication** - OTP & Email/Password login\n- 🛒 **Orders** - Complete order lifecycle\n- 📦 **Products** - Rich catalog\n- 💳 **Payments** - Razorpay integration\n- 🎁 **Promotions** - Gifts & Coupons\n- 👥 **Rooms** - Live shopping\n\n## Usage:\nRefer to 'API_INTEGRATION_GUIDE.md' for flow details.\n        ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    },
    {
      "url": "https://api.risbow.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter your JWT token from /auth/login",
        "in": "header"
      }
    },
    "schemas": {
...
      "UpdateProductDto": {
        "type": "object",
        "properties": {}
      },
...
      "RegisterVendorDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}